+++ flutter_custom_tabs_ios/ios/flutter_custom_tabs_ios/Sources/flutter_custom_tabs_ios/Launcher.swift
@@ -2,18 +2,46 @@ import Flutter
 import SafariServices
 import UIKit
 
-open class Launcher {
-  private var prewarmingTokenCache = [String: Any]()
+protocol Launcher {
+  func open(
+    _ url: URL,
+    options: [UIApplication.OpenExternalURLOptionsKey: Any],
+    completionHandler completion: ((Bool) -> Void)?
+  )
 
-  open func open(
+  func present(_ viewControllerToPresent: UIViewController, completion: ((Bool) -> Void)?)
+
+  func dismissAll(completion: (() -> Void)?)
+
+  func prewarmConnections(to urls: [URL]) -> String?
+
+  func invalidatePrewarmingSession(for sessionId: String)
+}
+
+extension Launcher {
+  func open(
     _ url: URL,
     options: [UIApplication.OpenExternalURLOptionsKey: Any] = [:],
     completionHandler completion: ((Bool) -> Void)? = nil
   ) {
+    open(url, options: options, completionHandler: completion)
+  }
+}
+
+// MARK: - Impl
+
+final class DefaultLauncher: Launcher {
+  private var prewarmingTokenCache = [String: Any]()
+
+  func open(
+    _ url: URL,
+    options: [UIApplication.OpenExternalURLOptionsKey: Any],
+    completionHandler completion: ((Bool) -> Void)?
+  ) {
     UIApplication.shared.open(url, options: options, completionHandler: completion)
   }
 
-  open func present(_ viewControllerToPresent: UIViewController, completion: ((Bool) -> Void)? = nil) {
+  func present(_ viewControllerToPresent: UIViewController, completion: ((Bool) -> Void)?) {
     guard let topViewController = UIWindow.keyWindow?.topViewController() else {
       completion?(false)
       return
@@ -23,7 +51,7 @@ open class Launcher {
     }
   }
 
-  open func dismissAll(completion: (() -> Void)? = nil) {
+  func dismissAll(completion: (() -> Void)?) {
     guard let rootViewController = UIWindow.keyWindow?.rootViewController else {
       completion?()
       return
@@ -44,7 +72,7 @@ open class Launcher {
     )
   }
 
-  open func prewarmConnections(to urls: [URL]) -> String? {
+  func prewarmConnections(to urls: [URL]) -> String? {
     guard #available(iOS 15.0, *) else {
       return nil
     }
@@ -55,7 +83,7 @@ open class Launcher {
     return id
   }
 
-  open func invalidatePrewarmingSession(withId sessionId: String) {
+  func invalidatePrewarmingSession(for sessionId: String) {
     guard #available(iOS 15.0, *) else {
       return
     }